"Added an rms calculation function.
This may be helpful for calculating grms. It can be imported (import
grms_calculator) and used as a function (grms_calculator.grms(arg) where
arg is a list of accelerometer readings. This can be implemented to
existing code, or used on data that has been previously stored.

Can we put some existing data we have acquired into this folder so I can
test this function?

-Dylan Frizzell
"