Added a power spectrum program
In the Accelerometer folder there is now two new files. These are grms.py and main.py. grms.py can calculate the rms of a set of data, take the fourier transform of the data, and display a normalized power spectrum. I am not 100% sure it is working correctly but I have used the pasco data also located in this folder to run some tests. To run the program as is type >python main.py <file_name>. As of right now the program is reading a set row of the csv file. I will add another command line arguement to select which accelerometer to read once we establish a convention.