Added data and finished the grms function
I added data I gathered in the repeatability study. This data can be analyzed with the main.py and grms.py scripts in the sample_pasco_data folder. The chuck and nochuck refers to the placement of the accelerometer. 'Chuck' means it was placed on the chuck. Otherwise it is bolted directly to the tabletop. The main function will provide a normalized power spectrum.