Added the datalogging feature.
The script 0803_6ch now has datalogging. I only added two import statements at the beginning, an exporter instance declaration, and an if statement in the update() method. The if statement take a mod 60 of a timer and rights a csv fle with each current time stamp. The mod 60 forces a 1 second export every 60 seconds. This can be changed as needed.