Create SteppingCode.txt
Started to outline the air regulator code.  The intro portion is what the pyserial generic example had us define at the beginning and I pulled in the values from our device.  The looping section is what I think the process should look like logically without syntax.  The minimalmodbus is a way to communicate with rs232 through pySerial.  You can see the example that shows how to change a register's value.  The ProportionAir sheet only gives me a "Starting Register" which I am guessing is the Registernumber that is called for in line 96.