Logic for air regulator cycle
Some of the syntax is not correct, but here is the logic of what I would like to have happening with the vibration cycling.  Here the user will be able to input the steplength and stepsize and then the code should account for that in that loop.  Can someone help me clean up the syntax?