Added the variable frequency code
There now exists two new files. varFreq.ino is currently running on the arduino that controls the cylinder mosfets. This code will run TWO cylinders and can be adapted to more. It constantly scans the serial port for a new timing variable.

The feqShift.py file is a class that has a changeFreq(int frequency) method. This method will send a new frequency to the arduino. Can be adapted to change the frequency on the fly.