Added the grms reading and writing features, fixed frequency shift
Hyoyeon and I added her code that reads the accelerometers and writes the grms
We added Writing.py that will read the accelerometers, calculate the grms, and write the grms to a text file.
Another function in Reading.py will read this text file for the cycling operation.
Everything should work as normal other than the checkGrms() method is not used and instead a
getGrms() method from Reading.py.

We also fixed a bug with the changeFreq method. Essentially adding a sleep 1 second timer
to allow the serial port to be established before the frequency can be altered.

We made minor changes to Vibration.py, varFreq.ino, and quickrun.py to accomodate our changes.